requirements.txt
torch
torchvision
onnx
onnxruntime
scikit-learn
matplotlib
numpy
pillow